requirements.txt
numpy==2.3.1
pandas==2.2.3
scipy==1.16.0
scikit-learn==1.7.2
statsmodels==0.14.6
torch==2.7.1
matplotlib==3.10.3
seaborn==0.13.2
pyyaml==6.0.2
openpyxl==3.1.5
